doc: update readme file
Simplified the package installation description, all previously
described complexities are not needed in most cases, and the current
command will work flawlessly on the listed version of go.

I removed the description of the examples, replaced them with links to
examples in the documentation so as not to duplicate and update several
places.