feat: add the ability to manually specify the want value
This is necessary to separate data in table tests when there are few
cases with big data and there are cases with short data that are better
represented in the table.

Perhaps this method is not a good idea, since the test data is mixed,
but perhaps this opportunity will do itself well in practice.