docs: update license file and headers
The copyright format has been changed to match the one specified on
https://spdx.org/licenses/BSD-3-Clause.html

Renamed the license file to match the copyright notice in the headers,
and this is also the more common name for the license file.

The license headers themselves were originally made following the
example of the open source repository by google.

Registration of a copyright notice in the README.md made in the style
of GOST R 7.0.1-2003