ci: update ci configuration
The previous configuration is badly outdated and needs to be updated.
In the new version, I removed the checks for the dependency option
without go modules, and also removed all old versions of go, as I don't
intend to support them in the future.

For convenience, added manual startup for workflow.